Rename nft -> sol_rl in training scripts and configs
- train_scripts/train_nft_*.py -> train_scripts/sol_rl/train_*.py
- configs/nft/ -> configs/sol_rl/
- Update all internal path references
